Move repository content to `._repository.py`
